Clean up subscene manager a bit
