+++ UnityProject/Assets/Scripts/Managers/SubSceneManager/SubSceneManager.SceneList.cs
@@ -1,0 +1,132 @@
+﻿using System.Collections;
+using Mirror;
+using UnityEditor;
+using UnityEngine.SceneManagement;
+
+//The scene list on the server
+public partial class SubSceneManager
+{
+	private string serverChosenAwaySite = "loading";
+	private string serverChosenMainStation = "loading";
+
+	public static string ServerChosenMainStation
+	{
+		get { return Instance.serverChosenMainStation; }
+	}
+
+	IEnumerator RoundStartServerLoadSequence()
+	{
+		var loadTimer = new SubsceneLoadTimer();
+		//calculate load time:
+		loadTimer.MaxLoadTime = 20f + (asteroidList.Asteroids.Count * 10f);
+		loadTimer.IncrementLoadBar("Preparing..");
+
+		//Choose and load a mainstation
+		yield return StartCoroutine(ServerLoadMainStation(loadTimer));
+		//Load Asteroids:
+		yield return StartCoroutine(ServerLoadAsteroids(loadTimer));
+		//Load away site:
+		yield return StartCoroutine(ServerLoadAwaySite(loadTimer));
+
+		netIdentity.isDirty = true;
+
+		yield return WaitFor.Seconds(0.1f);
+		UIManager.Display.preRoundWindow.CloseMapLoadingPanel();
+
+		Logger.Log($"Server has loaded {serverChosenAwaySite} away site", Category.SubScenes);
+	}
+
+	//Choose and load a main station on the server
+	IEnumerator ServerLoadMainStation(SubsceneLoadTimer loadTimer)
+	{
+		MainStationLoaded = true;
+		//Auto scene load stuff in editor:
+		var prevEditorScene = GetEditorPrevScene();
+		if (mainStationList.MainStations.Contains(prevEditorScene))
+		{
+			serverChosenMainStation = prevEditorScene;
+		}
+		else
+		{
+			serverChosenMainStation = mainStationList.GetRandomMainStation();
+		}
+		loadTimer.IncrementLoadBar($"Loading {serverChosenMainStation}");
+		//load main station
+		yield return StartCoroutine(LoadSubScene(serverChosenMainStation, loadTimer));
+		loadedScenesList.Add(new SceneInfo
+		{
+			SceneName = serverChosenMainStation,
+			SceneType = SceneType.MainStation
+		});
+	}
+
+	//Load all the asteroids on the server
+	IEnumerator ServerLoadAsteroids(SubsceneLoadTimer loadTimer)
+	{
+		loadTimer.IncrementLoadBar("Loading Asteroids");
+		foreach (var asteroid in asteroidList.Asteroids)
+		{
+			yield return StartCoroutine(LoadSubScene(asteroid, loadTimer));
+
+			loadedScenesList.Add(new SceneInfo
+			{
+				SceneName = asteroid,
+				SceneType = SceneType.Asteroid
+			});
+		}
+	}
+
+	//Load the away site on the server
+	IEnumerator ServerLoadAwaySite(SubsceneLoadTimer loadTimer)
+	{
+		var prevEditorScene = GetEditorPrevScene();
+		//Load the away site
+		if (awayWorldList.AwayWorlds.Contains(prevEditorScene))
+		{
+			serverChosenAwaySite = prevEditorScene;
+		}
+		else
+		{
+			serverChosenAwaySite = awayWorldList.GetRandomAwaySite();
+		}
+		loadTimer.IncrementLoadBar("Loading Away Site");
+		yield return StartCoroutine(LoadSubScene(serverChosenAwaySite, loadTimer));
+		AwaySiteLoaded = true;
+		loadedScenesList.Add(new SceneInfo
+		{
+			SceneName = serverChosenAwaySite,
+			SceneType = SceneType.AwaySite
+		});
+	}
+
+	string GetEditorPrevScene()
+	{
+		var prevEditorScene = "";
+#if UNITY_EDITOR
+		if (EditorPrefs.HasKey("prevEditorScene"))
+		{
+			if (!string.IsNullOrEmpty(EditorPrefs.GetString("prevEditorScene")))
+			{
+				prevEditorScene = EditorPrefs.GetString("prevEditorScene");
+			}
+		}
+#endif
+		return prevEditorScene;
+	}
+
+	/// <summary>
+	/// Add a new scene to a specific connections observable list
+	/// </summary>
+	void AddObservableSceneToConnection(NetworkConnection conn, Scene sceneContext)
+	{
+		if (!NetworkServer.observerSceneList.ContainsKey(conn))
+		{
+			AddNewObserverScenePermissions(conn);
+		}
+
+		if (!NetworkServer.observerSceneList[conn].Contains(sceneContext))
+		{
+			NetworkServer.observerSceneList[conn].Add(sceneContext);
+		}
+	}
+}
