+++ UnityProject/Assets/Scripts/Managers/SubSceneManager/SubSceneManager.SceneList.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: fe108d94bc61428498a0a500ee5050cd
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
